requirements.txt
streamlit
PyPDF2
openai
numpy
pandas
faiss-cpu
openpyxl

